fix: Remove OPEN_WEATHER key.

+++ app/main.py
@@ -15,46 +15,45 @@ app = FastAPI()
 
 load_dotenv()
 
-API_KEY = os.environ.get("OPEN_WEATHER_API_KEY", "28b2485cf1de8a48458937d968cc3ec9")
+# OpenWeather API Key
+API_KEY = os.environ.get("OPEN_WEATHER_API_KEY")
+if not API_KEY:
+    raise ValueError("API key not found")
 
-# MongoDB Configuration
-# MONGO_URL = os.getenv("MONGO_URL", "mongodb://mongo:27018")
-MONGO_URL = "mongodb://mongo"
+# Configuração do MongoDB
+MONGO_URL = os.getenv("MONGO_URL", "mongodb://mongo")
 client = MongoClient(MONGO_URL)
 db = client.weather_db
 
-# Get weather collection in MongoDB
 weather_collection = db.weather_collection
-
-# Get cities collection in MongoDB
 cities_collection = db.cities_collection
+
+# Ler as cidades do arquivo cities.txt e salvar no MongoDB
 cities = read_cities()
-# Delete all the existing cities
 cities_collection.delete_many({})
-# Insert the cities
 cities_collection.insert_one({"cities": cities})
 
 
-# Routes
+# Rotas
 @app.get("/weather-data")
 def get_weather_data():
-    # Get the weather data from MongoDB
+    # Recuperar os dados de temperatura e umidade do MongoDB
     data = weather_collection.find()
     data = [[item["id"], item["datetime"], item["weather_data"]] for item in data]
     if not data:
         return {"message": "Weather data not found"}
-    # Transform the data to a list
+    # Transformar os dados em uma lista
     return {"data": data, "message": "Weather data retrieved successfully"}
 
 
 @app.get("/cities")
 def get_cities():
-    # Get the cities from MongoDB
+    # Recuperar as cidades do MongoDB
     data = cities_collection.find()
     data = [item["cities"] for item in data]
     if not data:
         return {"message": "Cities not found"}
-    # Transform the data to a list
+    # Transformar os dados em uma lista
     return {"data": data, "message": "Cities retrieved successfully"}
 
 
@@ -68,16 +67,23 @@ async def fetch_weather(city_id: int):
 
 @app.post("/collect-weather/{request_id}")
 async def collect_weather(request_id: int):
-    # Collect weather data from Open Weather API and store it in MongoDB
+    # Coletar os dados de temperatura e umidade para cada cidade
     try:
-        # First, verify if the request_id is already in the database
+        # Primeiro, verificar se os dados já foram coletados
         if weather_collection.find_one({"id": request_id}):
             return {"message": "Weather data already collected for this request_id"}
+        # Extrair a hora atual
+        now = datetime.now()
         for idx, city_id in enumerate(cities):
+            if idx % 60 == 0:
+                # Calcule o tempo restante até o próximo minuto
+                current_second = now.second
+                remaining_seconds = 60 - current_second
+                await asyncio.sleep(remaining_seconds)
             data = await fetch_weather(city_id)
-            # Get the temperature and humidity from the data
+            # Extrair os dados necessários
             city_id, temperature, humidity = data["id"], data["main"]["temp"], data["main"]["humidity"]
-            # Store the data in MongoDB
+            # Salvar os dados no MongoDB
             weather_collection.insert_one({
                 "id": request_id,
                 "datetime": datetime.now(),
@@ -87,7 +93,6 @@ async def collect_weather(request_id: int):
                     "humidity": humidity
                 }
             })
-            await asyncio.sleep(1)
     except Exception as e:
         return {"message": f"Error while collecting weather data: {str(e)}"}
     return {"message": "Weather data collected successfully", "data": data}
@@ -95,10 +100,11 @@ async def collect_weather(request_id: int):
 
 @app.get("/get-collection-progress/{request_id}")
 def get_collection_progress(request_id: int) -> Dict[str, str]:
-    # Get the collection progress of the user from MongoDB
+    # Retornar o progresso da coleta de dados
     data = weather_collection.find({"id": request_id})
+    data = [item for item in data]
     if not data:
         return {"message": "Request ID not found"}
-    data = [item for item in data]
+    # Calcular o progresso da coleta
     percentage_progress = (len(data) / len(cities)) * 100
     return {"percentage_progress": f"{percentage_progress:.1f}%", "message": "Collection progress retrieved successfully"}
